django_backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
langdetect==1.0.7
mysqlclient==1.4.4
nltk==3.4.5
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
